copying initial code from tvldiff - compiles but does not run well

+++ scomm.go
@@ -3,22 +3,21 @@ package scomm
 import (
 	"bufio"
 	"errors"
-	"flag"
 	"fmt"
+	"io"
 	"log"
 	"math"
 	"os"
-	"runtime/pprof"
 	"strconv"
 	"strings"
 	"time"
 )
 
 var (
-	Profile, Verbose                   bool
-	SkipLines, BatchSize                                      int
+	Profile, Verbose                  bool
+	SkipLines, BatchSize              int
 	KeyParam, PayloadParam, Delimiter string
-	KeyPos,PayloadPos                             [][2]int
+	KeyPos, PayloadPos                [][2]int
 
 	cntLinesOld, cntLinesNew, cntSameLines, cntNewLines, updatedTags int
 	linesOld                                                         map[string]struct{}
@@ -27,9 +26,14 @@ var (
 )
 
 // TODO I dont think this is used anymore since KEY is also compound
-func getSimpleField(line string, pos [2]int, delimiter string) string {
-	if delimiter == "" {
+// getSimpleField extracts from a string line the substring or field defined by pos and optionally delimiter
+// pos in this case can look like either [4,7] => extract characters 4 to 7
+// or, with delimiter, [3,3]
+func getSimpleField(line string, pos [2]int, delim string) string {
+
+	if delim == "" {
 		var x, y int
+
 		if pos[0] == 0 {
 			x = 0
 		} else {
@@ -44,23 +48,25 @@ func getSimpleField(line string, pos [2]int, delimiter string) string {
 			log.Println("invalid data: " + line)
 			os.Exit(1)
 		}
-		return line[x:y]
-	}
 
-	ss := strings.Split(line, delimiter)
-	
-	if pos[0] > len(ss) {
-		log.Println("invalid data: " + line)
-		os.Exit(1)
-		// return ""
+		return line[x:y]
+	} else {
+		ss := strings.Split(line, delim)
+
+		if pos[0] > len(ss) {
+			log.Println("invalid data: " + line)
+			os.Exit(1)
+			// return ""
+		}
+
+		return ss[pos[0]-1]
 	}
-	return ss[pos[0]-1]
 }
 
-func getCompoundField(line string, pos [][2]int, delimiter string) string {
+func getCompoundField(line string, pos [][2]int, delim string) string {
 	var s string
 
-	if delimiter == "" {
+	if delim == "" { // position-based
 		for _, v := range pos {
 			var x, y int
 			if v[0] == 0 {
@@ -80,55 +86,56 @@ func getCompoundField(line string, pos [][2]int, delimiter string) string {
 			s += line[x:y]
 		}
 		return s
-	}
-
-	ss := strings.Split(line, delimiter)
-	for _, v := range pos {
-		if v[0] == v[1] { // single field
-			if v[0] > len(ss) {
-				log.Println("invalid data: " + line)
-				os.Exit(1)
-			}
-			s += ss[v[0]-1] + Delimiter
-		} else { // interval field like 3-7
-			if v[0] == 0 {
-				v[0] = 1
-			}
-			if v[1] == 0 {
-				v[1] = len(ss)
-			}
-			for w := v[0]; w <= v[1]; w++ {
-				if w > len(ss) {
+	} else { // field-based
+		ss := strings.Split(line, delim)
+		for _, v := range pos {
+			if v[0] == v[1] { // single field
+				if v[0] > len(ss) {
 					log.Println("invalid data: " + line)
 					os.Exit(1)
 				}
-				s += ss[w-1] + delimiter
+				s += ss[v[0]-1] + delim
+			} else { // interval field like 3-7
+				if v[0] == 0 {
+					v[0] = 1
+				}
+				if v[1] == 0 {
+					v[1] = len(ss)
+				}
+				for w := v[0]; w <= v[1]; w++ {
+					if w > len(ss) {
+						log.Println("invalid data: " + line)
+						os.Exit(1)
+					}
+					s += ss[w-1] + delim
+				}
 			}
 		}
+		// take out the last delimiter
+		return strings.TrimRight(s, delim)
 	}
-	// take out the last delimiter
-	return strings.TrimRight(s, delimiter)
 }
 
-// getTagLine returns all the tag related fields from the input line (TVL files may have extra info we do not use)
-func getCompoundFieldValue(line string, posKey, posPayload [][2]int, delimiter string) string {
-		return strings.Trim(getCompoundField(line, posKey, delimiter) + delimiter+ getCompoundField(line, posPayload, delimiter), delimiter)
-}
-
-// parseListItem parses an input int or int-int interval into an array [2]int
-func parseListItem(param string) (error, [2]int) {
+// parseListItem parses one input simple token (int or int-int or int- or -int) interval into an array [2]int
+// LIST = ITEM[,ITEM...]
+// ITEM = 3   => {3,3}
+//
+//	4-6 => {4,6}
+//	-7  => {0,7}
+//	8-  => {8,0}
+func parseItem(param string) ([2]int, error) {
 	var ret [2]int
 
 	if param == "" {
-		return errors.New("option requires range argument"), ret
+		return ret, errors.New("option requires range argument")
 	}
 
-	if strings.Contains(s, "-") {
+	if strings.Contains(param, "-") {
 
 		ss := strings.Split(param, "-")
 
 		if len(ss) > 2 {
-			return errors.New("invalid range " + param), ret
+			return ret, errors.New("invalid range " + param)
 		}
 
 		if ss[0] == "" {
@@ -136,90 +143,55 @@ func parseListItem(param string) (error, [2]int) {
 		} else {
 			i, err := strconv.Atoi(ss[0])
 			if err != nil {
-				return err, ret // TODO use fmt.Errorf
+				return ret, err // TODO use fmt.Errorf
 			}
 			ret[0] = i
 		}
 
 		if ss[1] == "" {
 			ret[1] = 0 // interval like "4-" which means "4-end of string"
-			return nil, ret
+			return ret, nil
 		}
 
 		i, err := strconv.Atoi(ss[1])
 		if err != nil {
-			return err, ret
+			return ret, err
 		}
 		if ret[0] != 0 && ret[0] > i {
-			return errors.New("reverted interval " + s), ret
+			return ret, errors.New("reverted interval " + param)
 		}
 		ret[1] = i
-		return nil, ret
-	}
+		return ret, nil
+	} else {
 
-	i, err := strconv.Atoi(param)
-	if err != nil {
-		return err, ret // TODO use fmt.Errorf
-	}
+		i, err := strconv.Atoi(param)
+		if err != nil {
+			return ret, err // TODO use fmt.Errorf
+		}
 
-	if Delimiter != "" && i <= 0 {
-		return errors.New("field is invalid " + param), ret
-	}
+		if i == 0 { // positions and fields are 1-based
+			return ret, errors.New("field is invalid " + param)
+		}
 
-	ret[0], ret[1] = i, i
-	return nil, ret
+		ret[0], ret[1] = i, i
+		return ret, nil
+	}
 }
 
-func parseList(param string) 
+func parseList(param string) ([][2]int, error) {
 
-// parseTagParams parses the agency, tag and payload parameters, which define the TVL structure; some projects may have extra data in the TVL file
-func parseTagParams(agencyParam, tagParam, payloadParam string) error {
-	var err error
+	var ret [][2]int
+	ss := strings.Split(param, ",")
 
-	if Delimiter != "" && strings.Contains(agencyParam, "-") {
-		return errors.New("cannot have list of fields: " + agencyParam)
-	}
-
-	err, agencyPos = parseListItem(agencyParam)
-	if err != nil {
-		return err
-	}
-
-	// if delimiter != "" && strings.Contains(tagParam, "-") {
-	// return errors.New("cannot have list of fields: " + tagParam)
-	// }
-
-	// err, tagPos = parseListItem(tagParam)
-	// if err != nil {
-	// 	return err
-	// }
-
-	ss := strings.Split(tagParam, ",")
 	for _, v := range ss {
-		err, y := parseListItem(v)
+		y, err := parseItem(v)
 		if err != nil {
-			return err
+			return ret, err
 		}
-		tagPos = append(tagPos, y)
+		ret = append(ret, y)
 	}
 
-	ss = strings.Split(payloadParam, ",")
-	for _, v := range ss {
-		err, y := parseListItem(v)
-		if err != nil {
-			return err
-		}
-		payloadPos = append(payloadPos, y)
-	}
-
-	if verbose {
-		log.Println("headerline", headerline)
-		log.Println("agency", agencyParam, agencyPos)
-		log.Println("tag", tagParam, tagPos)
-		log.Println("payload", payloadParam, payloadPos)
-	}
-
-	return nil
+	return ret, nil
 }
 
 //////////////////////////////////////////////////
@@ -229,42 +201,25 @@ func Scomm(
 	skipLines int,
 	keyParam string,
 	payloadParam string,
-	delimiter string,
+	dataDelim string,
 	batchSize int,
-	delimiterOut string, 
+	outputDelim string,
 	oldDataIn, newDataIn io.Reader,
-	newDataOut, oldDataOut, commDataOut io.Writer
+	newDataOut, oldDataOut, commDataOut io.Writer,
 ) error {
+
 	log.SetFlags(log.Ldate | log.Ltime)
 	log.Println("Start Scomm")
 
-	verbose("start scomm")
-	
-	if err := parseTagParams(); err != nil {
+	vrb("start scomm")
+
+	keyPos, err := parseList(keyParam)
+
+	if err != nil {
 		log.Println(err)
-		fmt.Println(err)
 		os.Exit(1)
 	}
-	
-	log.Println("test parameter: [", testParam, "]")
-	if testParam != "" {
-		ss := strings.Split(testParam, ":")
-		switch {
-		case testParam == "test1:default":
-			test1("1234567890abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ")
-			os.Exit(0)
-		case strings.Contains(testParam, "test1:"):
-			test1(ss[1])
-			os.Exit(0)
-		case testParam == "test2:default":
-			test2("1234,56789,0abcdef,ghijklmnopqrstu,vwxyzABC,DEFGHIJ,KLMNOP,QRST,U,V,WXYZ")
-			os.Exit(0)
-		case strings.Contains(testParam, "test2:"):
-			test2(ss[1])
-			os.Exit(0)
-		}
-	}
-	
+
 	ts1 := time.Now()
 	// if profile {
 	// 	proffile, err := os.Create("tvldiff.prof")
@@ -274,111 +229,83 @@ func Scomm(
 	// 	}
 	// 	pprof.StartCPUProfile(proffile)
 	// }
-	
+
 	// works with files only, no "Real" process substitution :((
-	// TODO make this a single function to call for each FD
-	file3, file3ok := getFDValid(3, "fdold")
+	file3, file3ok := IsFDValid(3, "oldDataIn")
 	if !file3ok {
-		os.Exit(1)
+		return errors.New("Bad file descriptor 3")
 	}
-	file4, file4ok := getFDValid(4, "fdnew")
+	file4, file4ok := IsFDValid(4, "newDataIn")
 	if !file4ok {
-		os.Exit(1)
+		return errors.New("Bad file descriptor 4")
 	}
-	file5, file5ok := getFDValid(5, "fdmerge")
+	file5, file5ok := IsFDValid(5, "newDataOut")
 	if !file5ok {
-		fmt.Println("Cannot write to fd5")
-		os.Exit(1)
+		return errors.New("Bad file descriptor 5")
 	}
-	file6, file6ok := getFDValid(6, "fddelete")
+	file6, file6ok := IsFDValid(6, "oldDataOut")
 	if !file6ok {
-		fmt.Println("Cannot write to fd6")
-		os.Exit(1)
+		return errors.New("Bad file descriptor 6")
 	}
-	file7, file7ok := getFDValid(7, "fdsame")
+	file7, file7ok := IsFDValid(7, "commonDataOut")
 	if !file7ok {
-		dbg("discard common lines")
+		return errors.New("Bad file descriptor 7")
 	}
-	
-	if testParam != "" {
-		switch {
-		case testParam == "test3":
-			sc := bufio.NewScanner(file3)
-			if sc.Scan() {
-				log.Println("read ok from file3")
-			} else {
-				log.Println("could not read from file3")
-				log.Println(sc.Err())
-			}
-			s := sc.Text()
-			log.Println(s)
-	
-			sc = bufio.NewScanner(file4)
-			if sc.Scan() {
-				log.Println("read ok from file4")
-			} else {
-				log.Println("could not read from file4")
-				log.Println(sc.Err())
-			}
-			s = sc.Text()
-			log.Println(s)
-	
-			os.Exit(0)
-		}
-	}
-	
+
 	var (
 		line     string
 		sc3, sc4 *bufio.Scanner
 		// headerNewDone  bool
 		// oldEOF, newEOF bool
 	)
-	
+
 	batchMode := batchSize > 0
 	// read OLD data from fd3 or stdin
 	sc3 = bufio.NewScanner(file3)
 	sc4 = bufio.NewScanner(file4)
-	
+
 	dbg("allocate memory")
 	if batchMode {
 		linesOld = make(map[string]struct{}, batchSize)
-		linesNew = make(map[string]struct{}, int(batchSize/100)) // I expect 1% tags to be new or updated
+		linesNew = make(map[string]struct{}, 2*int(batchSize/100)) // I expect 1-2% tags to be new or updated
 	} else {
 		linesOld = make(map[string]struct{})
 		linesNew = make(map[string]struct{})
 	}
-	newTagsList = make(map[string]struct{}) // ag+tag only from NEW file
-	
+	newKeysList = make(map[string]struct{})
+
 	if batchMode {
 		log.Println("start processing in batch mode, size", batchSize)
 	} else {
 		log.Println("start processing in full mode")
 	}
-	
+
 	// read both headers to get it over with
-	if headerline {
-		if sc3.Scan() {
-			log.Println("ignoring old tag data header line", sc3.Text())
-		} else {
-			// unable to even read one line, and header was specified - problem
-			log.Println("unable to read old header line")
-			fmt.Println("unable to read old header line")
-			os.Exit(1)
+	if skipLines > 0 {
+		for i := 1; i <= skipLines; i++ {
+			if sc3.Scan() {
+				log.Println("ignoring old tag data header line", sc3.Text())
+			} else {
+				// unable to even read one line, and header was specified - problem
+				log.Println("unable to read old header line")
+				os.Exit(1)
+			}
 		}
-		if sc4.Scan() {
-			log.Println("ignoring new tag data header line", sc4.Text())
-		} else {
-			// unable to even read one line, and header was specified - problem
-			log.Println("unable to read new header line")
-			fmt.Println("unable to read new header line")
-			os.Exit(1)
+		for i := 1; i <= skipLines; i++ {
+			if sc4.Scan() {
+				log.Println("ignoring new tag data header line", sc4.Text())
+			} else {
+				// unable to even read one line, and header was specified - problem
+				log.Println("unable to read new header line")
+				os.Exit(1)
+			}
 		}
 	}
-	
-	if batchMode { ///////////////////////////////////////////////////////////// batch mode
-	
+
+	if batchMode {
+
 		for { // read from OLD and NEW alternatively until both are done
-	
+
 			// read batchSize lines from OLD
 			for sc3.Scan() {
 				line = sc3.Text()
@@ -391,18 +318,18 @@ func Scomm(
 					break
 				}
 			}
-	
+
 			if err := sc3.Err(); err != nil {
 				log.Println("failed reading old tags:", err)
 				os.Exit(1)
 			}
 			log.Println("read", cntLinesOld, "old tags")
 			log.Println("old tags", len(linesOld), "new tags", len(linesNew), "matched tags", cntSameLines)
-	
+
 			// check existing linesNew, read in previous loop and not matched
 			for line, _ := range linesNew {
 				_, found := linesOld[line]
-	
+
 				if found { // same line exists in OLD, delete from OLD and do not add to NEW
 					cntSameLines++
 					delete(linesOld, line)
@@ -417,7 +344,7 @@ func Scomm(
 				}
 			}
 			log.Println("old buffer", len(linesOld), "new buffer", len(linesNew), "matched so far", cntSameLines)
-	
+
 			// read batchSize lines from NEW, and check against linesOld
 			for sc4.Scan() {
 				line = sc4.Text()
@@ -425,9 +352,9 @@ func Scomm(
 				if verbose && cntLinesNew%2_000_000 == 0 {
 					log.Println("read 2M new tags, total", cntLinesNew)
 				}
-	
+
 				_, found := linesOld[line]
-	
+
 				if found { // same line exists in OLD, delete from OLD and do not add to NEW
 					cntSameLines++
 					delete(linesOld, line)
@@ -440,32 +367,32 @@ func Scomm(
 					}
 				} else { // line does not exist in OLD, add to NEW
 					linesNew[line] = struct{}{}
-					newTagsList[getKey(line)] = struct{}{}
+					newKeysList[getCompoundField(line, keyPos, dataDelim)] = struct{}{}
 				}
 				if cntLinesNew%batchSize == 0 {
 					break
 				}
 			}
-	
+
 			if err := sc4.Err(); err != nil {
 				log.Println("failed reading from old tag file:", err)
 				os.Exit(1)
 			}
-	
+
 			log.Println("read", cntLinesNew, "new tags")
 			log.Println("old buffer", len(linesOld), "new buffer", len(linesNew), "matched", cntSameLines)
-	
+
 			if cntLinesOld%batchSize != 0 && cntLinesNew%batchSize != 0 {
 				break
 			}
 		} // read from OLD and NEW alternatively until both are done
-	
+
 		log.Println("read", cntLinesOld, "old tags,", cntLinesNew, "new tags,", cntSameLines, "matched,", cntNewLines, "preserved,")
-	
-	} else { /////////////////////////////////////////////////////////////////// full mode
-	
+
+	} else { // full mode
+
 		// read all OLD tags
-	
+
 		for sc3.Scan() {
 			line = sc3.Text()
 			cntLinesOld++
@@ -474,16 +401,16 @@ func Scomm(
 				log.Println("read 2M old tags, total", cntLinesOld)
 			}
 		}
-	
+
 		if err := sc3.Err(); err != nil {
 			log.Println("failed reading old tags:", err)
 			os.Exit(1)
 		}
-	
+
 		log.Println("read", cntLinesOld, "old tags,", len(linesOld), "are unique")
-	
+
 		// read all NEW tags
-	
+
 		for sc4.Scan() {
 			line = sc4.Text()
 			cntLinesNew++ // keep a count of lines read regardless if they existed in OLD
@@ -491,9 +418,9 @@ func Scomm(
 				log.Println("read 2M new tags, total", cntLinesNew)
 				log.Println("old tags", len(linesOld), "new tags", len(linesNew), "matched tags", cntSameLines)
 			}
-	
+
 			_, found := linesOld[line]
-	
+
 			if found { // same line exists in OLD, delete from OLD and do not add to NEW
 				cntSameLines++
 				delete(linesOld, line)
@@ -507,49 +434,47 @@ func Scomm(
 			} else { // line does not exist in OLD, add to NEW
 				cntNewLines++
 				linesNew[line] = struct{}{}
-				newTagsList[getKey(line)] = struct{}{}
+				if keyParam != "" {
+					newKeysList[getCompoundField(line, keyPos, dataDelim)] = struct{}{}
+				}
 			}
 		}
-	
+
 		if err := sc4.Err(); err != nil {
 			log.Println("failed reading new tags:", err)
 			os.Exit(1)
 		}
-	
+
 		log.Println("read", cntLinesOld, "old tags,", cntLinesNew, "new tags,", cntSameLines, "matched,", cntNewLines, "preserved,")
-	
+
 	} /////////////////////////////////////////////////////////////////// batch mode / full mode
-	
-	// looking now at agency+tag level
-	// tags in OLD that dont exist in NEW are deleted tags
-	// tags in NEW that dont exist in OLD are new tags
-	// tags in NEW that exist in OLD are UPDATED tags
-	
-	log.Println("searching for new and updated tags")
-	
-	for line, _ := range linesOld {
-		_, found := newTagsList[getKey(line)]
-		if found { // same ag+tg exists in NEW and OLD so it was changed, delete from OLD
-			updatedTags++
-			delete(linesOld, line)
+
+	if keyParam != "" {
+		// looking now at agency+tag level
+		// tags in OLD that dont exist in NEW are deleted tags
+		// tags in NEW that dont exist in OLD are new tags
+		// tags in NEW that exist in OLD are UPDATED tags
+
+		log.Println("searching for new and updated keys")
+
+		for line, _ := range linesOld {
+			_, found := newKeysList[getCompoundField(line, keyPos, dataDelim)]
+			if found { // same ag+tg exists in NEW and OLD so it was changed, delete from OLD
+				updatedTags++
+				delete(linesOld, line)
+			}
 		}
 	}
-	
+
 	s := fmt.Sprintf("new and updated tags: %d (%.2f%%), deleted tags: %d (%.2f%%)\n",
 		len(linesNew), float64(len(linesNew))*100/float64(cntLinesOld),
 		len(linesOld), float64(len(linesOld))*100/float64(cntLinesOld))
 	log.Println(s)
-	fmt.Println(s)
-	
-	// write the delete and merge files in batch mode
-	// for the new tag file, I have the full line
-	// but for the old tag, I may not have it - the data could come from an existing database table
-	// so for the delete file only write the ag+tg+pl
-	
-	log.Println("write merge and delete files")
-	
+
+	log.Println("write newDataOut and oldDataOut files")
+
 	done := make(chan error)
-	
+
 	go func() {
 		for str, _ := range linesNew {
 			_, err := file5.WriteString(str + "\n")
@@ -562,7 +487,7 @@ func Scomm(
 		log.Println("wrote merge file")
 		done <- nil
 	}()
-	
+
 	go func() {
 		for str, _ := range linesOld {
 			_, err := file6.WriteString(str + "\n")
@@ -575,50 +500,56 @@ func Scomm(
 		log.Println("wrote delete file")
 		done <- nil
 	}()
-	
-	exitcode := 0
-	if <-done != nil {
-		exitcode = 1
+
+	err = <-done
+
+	if err != nil {
+		log.Println(err)
+		<-done
+		return err
 	}
-	if <-done != nil {
-		exitcode = 1
+
+	err = <-done
+
+	if err != nil {
+		log.Println(err)
+		<-done
+		return err
 	}
-	
+
 	ts2 := time.Now()
-	if profile {
-		pprof.StopCPUProfile()
-	}
-	
-	log.Println("end tvldiff, time taken", math.Ceil(ts2.Sub(ts1).Seconds()), "sec")
-	os.Exit(exitcode)
-	}
+	// if profile {
+	// pprof.StopCPUProfile()
+	// }
+
+	log.Println("end scomm, time taken", math.Ceil(ts2.Sub(ts1).Seconds()), "sec")
+	return nil
+}
 
 /////////////
 
-	func verbose(params ...interface{}) {
-		if Verbose {
-			log.Println(params...)
-		}
+func vrb(params ...interface{}) {
+	if Verbose {
+		log.Println(params...)
 	}
-	
-/////////////
-	// TODO remove all calls to this or comment it out
-	func dbg(params ...interface{}) {
-			log.Println("DEBUG:", params...)
-	}
-	
+}
+
+// ///////////
+// TODO remove all calls to this or comment it out
+func dbg(params ...interface{}) {
+	log.Println(params...)
+}
+
 // IsFDValid returns a file from a file descriptor and if it ok to use
 func IsFDValid(fd int, name string) (*os.File, bool) {
 	f := os.NewFile(uintptr(fd), name)
 	if f == nil {
 		log.Println("invalid fd", fd, name)
-		fmt.Println("invalid fd", fd, name)
 		return f, false
 	}
 	_, err := f.Stat()
-	if err != nil && verbose {
+	if err != nil {
 		log.Println("cannot stat fd", fd, name)
-		fmt.Println("cannot stat fd", fd, name)
 	}
 	return f, err == nil
-	}
+}
